Create Code
Connects to the WildBerries (Russian Market place web), searches for all sellers of liquad soap and parces every product of the brand "SYNERGETIC".

It creates a database of each discovered product and organizes them into categories. With each code update, it adds new data to the table, reflecting changes in product positioning and ad performance. This data is then analyzed further.